Final restructuring pass
