+++ src/popup/vault/current-tab.component.ts
@@ -17,9 +17,8 @@ import { CipherView } from "jslib-common/models/view/cipherView";
 
 import { BrowserApi } from "../../browser/browserApi";
 import { AutofillService } from "../../services/abstractions/autofill.service";
+import { VaultFilterService } from "../../services/vaultFilter.service";
 import { PopupUtilsService } from "../services/popup-utils.service";
-
-import { VaultFilterService } from "./vault-filter/vault-filter.service";
 
 const BroadcasterSubscriptionId = "CurrentTabComponent";
