+++ src/popup/vault/vault-select.component.html
